refactor: Refactor core/commit and tests
